comparison does not appear automatically, but can be called for any of the possible solutions. cleaned up the code after fix, resulting in smaller classes
